requirements.txt
Django==5.0.2
djangorestframework==3.14.0
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1 